Updates to unit testing: Both Chance and Blackjack changed to test themselves instead of deck.
This caused all of the test for chance and black jack to fail. I suspect that the error that we were working on before in the unit testing for swap test is not just for swap. What ever the problem is is effecting all the test that call on the Mocks.